chore: switch to PasswordInput component
