+++ src/ui/pages/Login/login.scss
@@ -4,7 +4,7 @@
   color: $primary-green;
 
   &__content {
-    .icon {
+    .login-icon {
       width: 8rem !important;
       height: 8rem !important;
     }
