+++ src/ui/pages/Login/useLogin.ts
@@ -1,9 +1,9 @@
 import { BaseSyntheticEvent, useCallback } from "react";
 import { UseFormRegister, useForm } from "react-hook-form";
 
+import { PasswordFormFields } from "@src/types";
 import { unlock } from "@src/ui/ducks/app";
 import { useAppDispatch } from "@src/ui/ducks/hooks";
-import { PasswordFormFields } from "@src/types";
 
 export interface IUseLoginData {
   isLoading: boolean;
