+++ src/ui/pages/Onboarding/Onboarding.tsx
@@ -1,12 +1,7 @@
-import { InputAdornment } from "@mui/material";
-import Tooltip from "@mui/material/Tooltip";
-
 import logoSVG from "@src/static/icons/logo.svg";
 import { ButtonType, Button } from "@src/ui/components/Button";
 import { Icon } from "@src/ui/components/Icon";
-import { Input } from "@src/ui/components/Input";
 import { PasswordInput } from "@src/ui/components/PasswordInput";
-import { useShowPassword } from "@src/ui/hooks/showPassword";
 
 import "./onboarding.scss";
 import { useOnboarding } from "./useOnboarding";
@@ -25,12 +20,7 @@ const Onboarding = (): JSX.Element => {
 
         <div className="text-base">To continue, please setup a password</div>
 
-        <PasswordInput
-          isShowInfo={true}
-          isShowConfirmPassword={true}
-          register={register}
-          errors={errors}
-        />
+        <PasswordInput isShowConfirmPassword isShowInfo errors={errors} register={register} />
       </div>
 
       {errors.root && <div className="text-red-500 text-sm text-center">{errors.root}</div>}
