+++ src/ui/pages/Onboarding/useOnboarding.ts
@@ -2,10 +2,10 @@ import { BaseSyntheticEvent, useCallback } from "react";
 import { useForm, UseFormRegister } from "react-hook-form";
 import { object, ref, string } from "yup";
 
+import { PasswordFormFields } from "@src/types";
 import { setupPassword } from "@src/ui/ducks/app";
 import { useAppDispatch } from "@src/ui/ducks/hooks";
 import { useValidationResolver } from "@src/ui/hooks/validation";
-import { PasswordFormFields } from "@src/types";
 
 export interface IUseOnboardingData {
   isLoading: boolean;
